+++ change-log.md
@@ -11,6 +11,7 @@ Improvements:
     - `DI\autowire()` autowires an object and allows to override specific constructor and method parameters
 - [#494](https://github.com/PHP-DI/PHP-DI/pull/494) The container can now be compiled for optimum performances in production
 - The container can now be built without parameters: `new Container()`
+- [#490](https://github.com/PHP-DI/PHP-DI/issues/490) Support definitions in nested arrays (by [@yuloh](https://github.com/yuloh))
 - [#242](https://github.com/PHP-DI/PHP-DI/issues/242) Error in case a definition is not indexed by a string
 
 BC breaks: